32-bit fix

+++ pypy/module/cpyext/test/test_bytearrayobject.py
@@ -30,7 +30,8 @@ class AppTestStringObject(AppTestCpythonExtensionBase):
                  #endif
                  if(s->ob_type->tp_basicsize != expected_size)
                  {
-                     printf("tp_basicsize==%ld\\n", s->ob_type->tp_basicsize); 
+                     printf("tp_basicsize==%ld\\n",
+                            (long)s->ob_type->tp_basicsize); 
                      result = 0;
                  }
                  Py_DECREF(s);
